backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic[email]==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
alembic==1.14.1
weasyprint==63.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
aiosqlite==0.20.0
